api/requirements.txt
adalflow==1.1.3 ; python_version >= "3.11" and python_version < "4.0"
aiohappyeyeballs==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.13.1 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.11.0 ; python_version >= "3.11" and python_version < "4.0"
asyncstdlib==3.13.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==25.4.0 ; python_version >= "3.11" and python_version < "4.0"
azure-core==1.36.0 ; python_version >= "3.11" and python_version < "4.0"
azure-identity==1.25.1 ; python_version >= "3.11" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.40.56 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.40.56 ; python_version >= "3.11" and python_version < "4.0"
cachetools==6.2.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.10.5 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and platform_python_implementation != "PyPy" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_version < "4.0"
click==8.3.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0"
cryptography==46.0.3 ; python_version >= "3.11" and python_version < "4.0"
diskcache==5.6.3 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
faiss-cpu==1.11.0.post1 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.119.1 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
google-ai-generativelanguage==0.6.15 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.26.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.26.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-python-client==2.185.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.41.1 ; python_version >= "3.11" and python_version < "4.0"
google-generativeai==0.8.5 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.71.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.71.2 ; python_version >= "3.11" and python_version < "3.14"
grpcio==1.76.0 ; python_version >= "3.11" and python_version < "3.14"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.31.0 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.11.1 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
jsonlines==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
langid==1.1.6 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
msal-extensions==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
msal==1.34.0 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.7.0 ; python_version >= "3.11" and python_version < "4.0"
nest-asyncio==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.3.4 ; python_version >= "3.11" and python_version < "4.0"
ollama==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
openai==2.6.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.5 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.11" and platform_python_implementation != "PyPy" and python_version < "4.0" and implementation_name != "PyPy"
pydantic-core==2.41.4 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.12.3 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.11" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.2.5 ; python_version >= "3.11" and python_version < "4.0"
pytest-asyncio==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.4.2 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.11" and python_version < "4.0"
regex==2025.10.23 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4"
s3transfer==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.48.0 ; python_version >= "3.11" and python_version < "4.0"
tiktoken==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn[standard]==0.38.0 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.22.1 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "4.0"
watchfiles==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.22.0 ; python_version >= "3.11" and python_version < "4.0"
